Module locale::facet

Infrastructure for defining facets.

Facet is an object that defines data and logic required for some aspect (category) of localization. For example Numeric facet deals with formatting and parsing numbers, Time facet deals with formatting and parsing dates and times, Messages deals with translated messages and so on.

To use facets, all you need is the convenience interface on locale::Locale. If that's what you need, you can stop reading now.

But if you want to define your own facets, this module provides common infrastructure for caching facet instances, defining a default implementations and default factory functions to construct them for recognized locale and making sure a fall-back "invariant" implementation is available if nothing better can be obtained.

Structs

Builder
Tag type on which the Factory is implemented.

Traits

Factory
Trait defining how a facet is to be instantiated by default.

Functions

get
Get instance of facet for given Locale.

try_get_by_tag
Get instance of facet for given LanguageRange, it it exists.