Function locale::facet::try_get_by_tag

pub fn try_get_by_tag<'a, F: Any + Send + Sync + ?Sized>(t: &LanguageRange<'a>)
                                                         -> Option<Arc<F>> where Builder: Factory<F>

Get instance of facet for given LanguageRange, it it exists.

This is an auxiliary function intended mainly for facet factories if the facet should depend on another facet for the same LanguageRange.

The instance is constructed if needed, and cached for reuse. If the corresponding factory does not understand the tag, None is returned, so the search can properly continue with any fallback option.

Note, however, that the factory might return instance that only matches partially as long as it understood at least the language.