Struct locale::facet::Builder

pub struct Builder;

Tag type on which the Factory is implemented.

We need to specialize factory functions for each facet. Since Rust does not have explicit specializations, we emulate them by implementing for this tag type.