Trait locale::facet::Factory

pub trait Factory<F: Any + Send + Sync + ?Sized> {
    fn category() -> &'static str;
    fn new_for<'a>(_: &LanguageRange<'a>) -> Option<Arc<F>>;
    fn new_invariant() -> Arc<F>;
}

Trait defining how a facet is to be instantiated by default.

To define a facet, this trait, specialized to the facet trait, must be implemented for the Builder tag type.

Required Methods

fn category() -> &'static str

Name of category this trait implements.

This should be an associated constant, but those are not stable yet.

fn new_for<'a>(_: &LanguageRange<'a>) -> Option<Arc<F>>

Construct facet for given LanguageRange

If the tag is not understood or no data are available for it, this functions should return None. The get function will then take care of trying the fallback tags and, if that fails too, uses new_invariant() to get invariant implementation.

fn new_invariant() -> Arc<F>

Construct facet for invariant locale.

Every facet must have an invariant implementation. That will be used as last resort if there is no other implementation for any of the users locales (or if user requests the invariant locale).