Module locale::linux::ffi

Bindings for GNU LibC localization functions.

This was generated by rust-bindgen from <locale.h>, <langinfo.h> and <iconv.h>. Iconv is included for converting output from nl_langinfo_l to utf-8 in case the user-selected locale is not utf-8. Thankfully in Linux all these functions are part of the libc itself, so they are already available and we don't need to do any additional linking and so there is no need to put this in separate crate either.

Structs

Struct___locale_struct

Struct_lconv

Enums

Struct___locale_data

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

ALT_DIGITS

AM_STR

CODESET

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

D_FMT

D_T_FMT

ERA

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

LC_ADDRESS

LC_ADDRESS_MASK

LC_ALL

LC_ALL_MASK

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_IDENTIFICATION

LC_IDENTIFICATION_MASK

LC_MEASUREMENT

LC_MEASUREMENT_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NAME

LC_NAME_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_PAPER

LC_PAPER_MASK

LC_TELEPHONE

LC_TELEPHONE_MASK

LC_TIME

LC_TIME_MASK

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

PM_STR

RADIXCHAR

THOUSEP

T_FMT

T_FMT_AMPM

_DATE_FMT

_NL_ADDRESS_CODESET

_NL_ADDRESS_COUNTRY_AB2

_NL_ADDRESS_COUNTRY_AB3

_NL_ADDRESS_COUNTRY_CAR

_NL_ADDRESS_COUNTRY_ISBN

_NL_ADDRESS_COUNTRY_NAME

_NL_ADDRESS_COUNTRY_NUM

_NL_ADDRESS_COUNTRY_POST

_NL_ADDRESS_LANG_AB

_NL_ADDRESS_LANG_LIB

_NL_ADDRESS_LANG_NAME

_NL_ADDRESS_LANG_TERM

_NL_ADDRESS_POSTAL_FMT

_NL_COLLATE_CODESET

_NL_COLLATE_COLLSEQMB

_NL_COLLATE_COLLSEQWC

_NL_COLLATE_EXTRAMB

_NL_COLLATE_EXTRAWC

_NL_COLLATE_GAP1

_NL_COLLATE_GAP2

_NL_COLLATE_GAP3

_NL_COLLATE_INDIRECTMB

_NL_COLLATE_INDIRECTWC

_NL_COLLATE_NRULES

_NL_COLLATE_RULESETS

_NL_COLLATE_SYMB_EXTRAMB

_NL_COLLATE_SYMB_HASH_SIZEMB

_NL_COLLATE_SYMB_TABLEMB

_NL_COLLATE_TABLEMB

_NL_COLLATE_TABLEWC

_NL_COLLATE_WEIGHTMB

_NL_COLLATE_WEIGHTWC

_NL_CTYPE_CLASS

_NL_CTYPE_CLASS32

_NL_CTYPE_CLASS_NAMES

_NL_CTYPE_CLASS_OFFSET

_NL_CTYPE_CODESET_NAME

_NL_CTYPE_EXTRA_MAP_1

_NL_CTYPE_EXTRA_MAP_2

_NL_CTYPE_EXTRA_MAP_3

_NL_CTYPE_EXTRA_MAP_4

_NL_CTYPE_EXTRA_MAP_5

_NL_CTYPE_EXTRA_MAP_6

_NL_CTYPE_EXTRA_MAP_7

_NL_CTYPE_EXTRA_MAP_8

_NL_CTYPE_EXTRA_MAP_9

_NL_CTYPE_EXTRA_MAP_10

_NL_CTYPE_EXTRA_MAP_11

_NL_CTYPE_EXTRA_MAP_12

_NL_CTYPE_EXTRA_MAP_13

_NL_CTYPE_EXTRA_MAP_14

_NL_CTYPE_GAP1

_NL_CTYPE_GAP2

_NL_CTYPE_GAP3

_NL_CTYPE_GAP4

_NL_CTYPE_GAP5

_NL_CTYPE_GAP6

_NL_CTYPE_INDIGITS0_MB

_NL_CTYPE_INDIGITS0_WC

_NL_CTYPE_INDIGITS1_MB

_NL_CTYPE_INDIGITS1_WC

_NL_CTYPE_INDIGITS2_MB

_NL_CTYPE_INDIGITS2_WC

_NL_CTYPE_INDIGITS3_MB

_NL_CTYPE_INDIGITS3_WC

_NL_CTYPE_INDIGITS4_MB

_NL_CTYPE_INDIGITS4_WC

_NL_CTYPE_INDIGITS5_MB

_NL_CTYPE_INDIGITS5_WC

_NL_CTYPE_INDIGITS6_MB

_NL_CTYPE_INDIGITS6_WC

_NL_CTYPE_INDIGITS7_MB

_NL_CTYPE_INDIGITS7_WC

_NL_CTYPE_INDIGITS8_MB

_NL_CTYPE_INDIGITS8_WC

_NL_CTYPE_INDIGITS9_MB

_NL_CTYPE_INDIGITS9_WC

_NL_CTYPE_INDIGITS_MB_LEN

_NL_CTYPE_INDIGITS_WC_LEN

_NL_CTYPE_MAP_NAMES

_NL_CTYPE_MAP_OFFSET

_NL_CTYPE_MAP_TO_NONASCII

_NL_CTYPE_MB_CUR_MAX

_NL_CTYPE_NONASCII_CASE

_NL_CTYPE_OUTDIGIT0_MB

_NL_CTYPE_OUTDIGIT0_WC

_NL_CTYPE_OUTDIGIT1_MB

_NL_CTYPE_OUTDIGIT1_WC

_NL_CTYPE_OUTDIGIT2_MB

_NL_CTYPE_OUTDIGIT2_WC

_NL_CTYPE_OUTDIGIT3_MB

_NL_CTYPE_OUTDIGIT3_WC

_NL_CTYPE_OUTDIGIT4_MB

_NL_CTYPE_OUTDIGIT4_WC

_NL_CTYPE_OUTDIGIT5_MB

_NL_CTYPE_OUTDIGIT5_WC

_NL_CTYPE_OUTDIGIT6_MB

_NL_CTYPE_OUTDIGIT6_WC

_NL_CTYPE_OUTDIGIT7_MB

_NL_CTYPE_OUTDIGIT7_WC

_NL_CTYPE_OUTDIGIT8_MB

_NL_CTYPE_OUTDIGIT8_WC

_NL_CTYPE_OUTDIGIT9_MB

_NL_CTYPE_OUTDIGIT9_WC

_NL_CTYPE_TOLOWER

_NL_CTYPE_TOLOWER32

_NL_CTYPE_TOUPPER

_NL_CTYPE_TOUPPER32

_NL_CTYPE_TRANSLIT_DEFAULT_MISSING

_NL_CTYPE_TRANSLIT_DEFAULT_MISSING_LEN

_NL_CTYPE_TRANSLIT_FROM_IDX

_NL_CTYPE_TRANSLIT_FROM_TBL

_NL_CTYPE_TRANSLIT_IGNORE

_NL_CTYPE_TRANSLIT_IGNORE_LEN

_NL_CTYPE_TRANSLIT_TAB_SIZE

_NL_CTYPE_TRANSLIT_TO_IDX

_NL_CTYPE_TRANSLIT_TO_TBL

_NL_CTYPE_WIDTH

_NL_IDENTIFICATION_ABBREVIATION

_NL_IDENTIFICATION_ADDRESS

_NL_IDENTIFICATION_APPLICATION

_NL_IDENTIFICATION_AUDIENCE

_NL_IDENTIFICATION_CATEGORY

_NL_IDENTIFICATION_CODESET

_NL_IDENTIFICATION_CONTACT

_NL_IDENTIFICATION_DATE

_NL_IDENTIFICATION_EMAIL

_NL_IDENTIFICATION_FAX

_NL_IDENTIFICATION_LANGUAGE

_NL_IDENTIFICATION_REVISION

_NL_IDENTIFICATION_SOURCE

_NL_IDENTIFICATION_TEL

_NL_IDENTIFICATION_TERRITORY

_NL_IDENTIFICATION_TITLE

_NL_MEASUREMENT_CODESET

_NL_MEASUREMENT_MEASUREMENT

_NL_MESSAGES_CODESET

_NL_MONETARY_CODESET

_NL_MONETARY_CONVERSION_RATE

_NL_MONETARY_CRNCYSTR

_NL_MONETARY_DECIMAL_POINT_WC

_NL_MONETARY_DUO_CURRENCY_SYMBOL

_NL_MONETARY_DUO_FRAC_DIGITS

_NL_MONETARY_DUO_INT_CURR_SYMBOL

_NL_MONETARY_DUO_INT_FRAC_DIGITS

_NL_MONETARY_DUO_INT_N_CS_PRECEDES

_NL_MONETARY_DUO_INT_N_SEP_BY_SPACE

_NL_MONETARY_DUO_INT_N_SIGN_POSN

_NL_MONETARY_DUO_INT_P_CS_PRECEDES

_NL_MONETARY_DUO_INT_P_SEP_BY_SPACE

_NL_MONETARY_DUO_INT_P_SIGN_POSN

_NL_MONETARY_DUO_N_CS_PRECEDES

_NL_MONETARY_DUO_N_SEP_BY_SPACE

_NL_MONETARY_DUO_N_SIGN_POSN

_NL_MONETARY_DUO_P_CS_PRECEDES

_NL_MONETARY_DUO_P_SEP_BY_SPACE

_NL_MONETARY_DUO_P_SIGN_POSN

_NL_MONETARY_DUO_VALID_FROM

_NL_MONETARY_DUO_VALID_TO

_NL_MONETARY_THOUSANDS_SEP_WC

_NL_MONETARY_UNO_VALID_FROM

_NL_MONETARY_UNO_VALID_TO

_NL_NAME_CODESET

_NL_NAME_NAME_FMT

_NL_NAME_NAME_GEN

_NL_NAME_NAME_MISS

_NL_NAME_NAME_MR

_NL_NAME_NAME_MRS

_NL_NAME_NAME_MS

_NL_NUM

_NL_NUMERIC_CODESET

_NL_NUMERIC_DECIMAL_POINT_WC

_NL_NUMERIC_THOUSANDS_SEP_WC

_NL_NUM_LC_ADDRESS

_NL_NUM_LC_COLLATE

_NL_NUM_LC_CTYPE

_NL_NUM_LC_IDENTIFICATION

_NL_NUM_LC_MEASUREMENT

_NL_NUM_LC_MESSAGES

_NL_NUM_LC_MONETARY

_NL_NUM_LC_NAME

_NL_NUM_LC_NUMERIC

_NL_NUM_LC_PAPER

_NL_NUM_LC_TELEPHONE

_NL_NUM_LC_TIME

_NL_PAPER_CODESET

_NL_PAPER_HEIGHT

_NL_PAPER_WIDTH

_NL_TELEPHONE_CODESET

_NL_TELEPHONE_INT_PREFIX

_NL_TELEPHONE_INT_SELECT

_NL_TELEPHONE_TEL_DOM_FMT

_NL_TELEPHONE_TEL_INT_FMT

_NL_TIME_CAL_DIRECTION

_NL_TIME_CODESET

_NL_TIME_ERA_ENTRIES

_NL_TIME_ERA_NUM_ENTRIES

_NL_TIME_FIRST_WEEKDAY

_NL_TIME_FIRST_WORKDAY

_NL_TIME_TIMEZONE

_NL_TIME_WEEK_1STDAY

_NL_TIME_WEEK_1STWEEK

_NL_TIME_WEEK_NDAYS

_NL_WABDAY_1

_NL_WABDAY_2

_NL_WABDAY_3

_NL_WABDAY_4

_NL_WABDAY_5

_NL_WABDAY_6

_NL_WABDAY_7

_NL_WABMON_1

_NL_WABMON_2

_NL_WABMON_3

_NL_WABMON_4

_NL_WABMON_5

_NL_WABMON_6

_NL_WABMON_7

_NL_WABMON_8

_NL_WABMON_9

_NL_WABMON_10

_NL_WABMON_11

_NL_WABMON_12

_NL_WALT_DIGITS

_NL_WAM_STR

_NL_WDAY_1

_NL_WDAY_2

_NL_WDAY_3

_NL_WDAY_4

_NL_WDAY_5

_NL_WDAY_6

_NL_WDAY_7

_NL_WD_FMT

_NL_WD_T_FMT

_NL_WERA_D_FMT

_NL_WERA_D_T_FMT

_NL_WERA_T_FMT

_NL_WERA_YEAR

_NL_WMON_1

_NL_WMON_2

_NL_WMON_3

_NL_WMON_4

_NL_WMON_5

_NL_WMON_6

_NL_WMON_7

_NL_WMON_8

_NL_WMON_9

_NL_WMON_10

_NL_WMON_11

_NL_WMON_12

_NL_WPM_STR

_NL_WT_FMT

_NL_WT_FMT_AMPM

_NL_W_DATE_FMT

__CURRENCY_SYMBOL

__DECIMAL_POINT

__ERA_YEAR

__FRAC_DIGITS

__GROUPING

__INT_CURR_SYMBOL

__INT_FRAC_DIGITS

__INT_N_CS_PRECEDES

__INT_N_SEP_BY_SPACE

__INT_N_SIGN_POSN

__INT_P_CS_PRECEDES

__INT_P_SEP_BY_SPACE

__INT_P_SIGN_POSN

__MON_DECIMAL_POINT

__MON_GROUPING

__MON_THOUSANDS_SEP

__NEGATIVE_SIGN

__NOEXPR

__NOSTR

__N_CS_PRECEDES

__N_SEP_BY_SPACE

__N_SIGN_POSN

__POSITIVE_SIGN

__P_CS_PRECEDES

__P_SEP_BY_SPACE

__P_SIGN_POSN

__THOUSANDS_SEP

__YESEXPR

__YESSTR

Functions

catclose⚠

catgets⚠

catopen⚠

duplocale⚠

freelocale⚠

iconv⚠

iconv_close⚠

iconv_open⚠

localeconv⚠

newlocale⚠

nl_langinfo⚠

nl_langinfo_l⚠

setlocale⚠

uselocale⚠

Type Definitions

Enum_Unnamed1

__locale_t

iconv_t

locale_t

nl_catd

nl_item

size_t