Struct locale::linux::IConv

pub struct IConv { /* fields omitted */ }

Methods

impl IConv

Wrapper for iconv.

See iconv(3).

On Linux this is part of standard C library and should always be able to convert any charset that the locale component presents, so we can conveniently use it for translating that to the Rust's internal utf-8 encoding there.

fn new(to: &str, from: &str) -> Result<Self>

Construct iconv converter.

See iconv_open(3).

fn convert(&self, src: &[u8], dst: &mut [u8]) -> (isize, usize, usize)

Convert data with iconv

See iconv(3). The parameters are

src: The input buffer.

dst: The output buffer.

Return values are:

Result of iconv. If -1, the reason can be read from errno (unfortunately ::std::io::Error::last_os_error() does not seem to be able to distinguish them at the moment).

Number of bytes processed from src.

Number of bytes written to dst.

The C interface returns the remaining buffers instead, but that is actually hard to work with in Rust.