Module locale::numeric

Numeric category.

Handling of local variations in how numbers are written and read.

Supported features

The default implementation currently only supports formatting numbers in decimal notation.

It is possible to provide an enhanced implementation in a separate crate.

Format specifications

The general form of format specifier is:

format_spec ::= [[fill]align][sign]["0"][width][.["0"]precision][type]
fill        ::= <any character>
align       ::= "<" | ">" | "=" | "^"
sign        ::= "+" | "-" | " " | "|"
width       ::= integer
precision   ::= integer
type        ::= depends on argument type

The fill and alignment options are:

<: Align to the logical left.

>: Align to the logical right.

=: Insert padding between sign and magnitude.

^: Centre the field with bias to the logical right.

The sign options are:

+: Print + for positive and - for negative numbers.

-: Print nothing for positive and - for negative numbers.

space: Print non-breakable space for positive and - for negative numbers.

|: Do not print sign at all (intended mainly as helper for formatting more complex types).

width is positive integer defining minimum width of the field. The field will never be truncated. The width is counted roughly in “symbols”, that is each digit, separator, sign and fill character count as 1.

If the width starts with 0, padding will be done by increasing number of integral digits displayed. The added digits will be zeroes from appropriate digit set and will have separators inserted between them.

The precision specifies number of digits after decimal point for f format and number of significant digits for the other floating-point formats. It has no effect for integers.

If the precision starts with 0, trailing zeroes after decimal point will be truncated.

Integer format types

n: The default format according to current locale. Can be omitted.

c: Use locale digits, but don't insert group separators.

C: Not localized. Use latin digits and don't insert group separators.

Floating-point number format types

e: Engineering exponential format. precision indicates the number of significant digits. Short exponent separator will be used which is E in most locales.

E: Common exponential format. precision indicates the number of significant digits. Long exponent separator will be used which is usually ×10 and in locales using the latin digits the superscript digits will be used to render the exponent.

f: Fixed point representation. precision indicates number of digits after decimal point.

g: General representation. precision indicates number of significant digits. Unlike C, the format will not switch to exponential representation for too small or too large numbers.

h: General representation with exponential fallback. This will switch to engineering exponential format if insignifiant zeroes would be otherwise needed, that is when the number of digits before decimal point would be larger than number of significant digits or if the first significant digit is lower order than tenths.

H: General representation with exponential fallback. Like h, but will fall back to common exponential format like E.

m: Mantissa. Just the mantissa part of the exponential format. For when something needs to be inserted between mantissa and exponent.

x: Exponent. Just the exponent part of the exponential format. For when something needs to be inserted between mantissa and exponent.

X: Exponent, but in locales with latin digits it will be in superscript.

Default floating-point format is f and default precision is .03.

Examples

Traits

ExpDisplay
Auxiliary trait for making trait objects from floating point numbers.

Numeric
Interface for formatting numbers.