Trait locale::numeric::Numeric

pub trait Numeric: Any + Send + Sync {
    fn format_int_to(&self,
                     n: &Display,
                     fmt: &str,
                     out: &mut Formatter)
                     -> Result { ... }
    fn format_float_to(&self,
                       n: &ExpDisplay,
                       fmt: &str,
                       out: &mut Formatter)
                       -> Result { ... }
    fn decimal_separator(&self) -> &str { ... }
    fn group_separator(&self) -> &str { ... }
    fn grouping(&self) -> &[u8] { ... }
    fn min_grouping_digits(&self) -> i32 { ... }
    fn decimal_digits(&self) -> &str { ... }
    fn plus_sign(&self) -> &str { ... }
    fn minus_sign(&self) -> &str { ... }
    fn engineering_exponent_separator(&self) -> &str { ... }
    fn common_exponent_separator(&self) -> &str { ... }
    fn infinity(&self) -> &str { ... }
    fn not_a_number(&self) -> &str { ... }
}

Interface for formatting numbers.

Currently only decimal notation is supported.

Default implementation

All methods have default implementation suitable for the invariant locale.

TODO

Parsing.

Rule-based formatting (Roman numerals, words and other non-decimal systems).

Abbreviated formats (“x mi.”, “xy bn.” etc.).

Provided Methods

fn format_int_to(&self, n: &Display, fmt: &str, out: &mut Formatter) -> Result

Format integer.

Output will be written to a formatter.

Note that padding is defined in fmt. Currently the padding in out is ignored, but in future it might be added to the one in fmt, if appropriate.

Parameters

n: The number to format. It's Display should produce string of ASCII digits, possibly preceded by '-' sign.

fmt: Format specification.

out: Output sink.

Errors

Returns fmt::Error on invalid patter or if the underlying formatter returns it during formatting.

fn format_float_to(&self,
n: &ExpDisplay,
fmt: &str,
out: &mut Formatter)
-> Result

Format floating point number.

Output will be written to a formatter.

Note that padding is defined in fmt. Currently the padding in out is ignored, but in future it might be added to the one in fmt, if appropriate.

Parameters

n: The number to format. It's Display and LowerExp should behave like those of built-in floating types.

fmt: Format specification. See DecimalFloatFmt for format description.

out: Output sink.

Errors

Returns fmt::Error on invalid patter or if the underlying formatter returns it during formatting.

fn decimal_separator(&self) -> &str

Decimal separator

fn group_separator(&self) -> &str

Thousands separator

Empty string if not used.

fn grouping(&self) -> &[u8]

Grouping

Primary group, that is the one including units, is first. The last group is used for all larger orders. A group with size 0 means no further group separators should be inserted.

fn min_grouping_digits(&self) -> i32

Mininimum grouping digits

Avoid creating groups of less than this size. E.g. if 2, format 10,000, but 1000, not 1,000.

fn decimal_digits(&self) -> &str

String of decimal digits

String of consecutive decimal digits 0 to 9. The digits must come from the same digit set.

fn plus_sign(&self) -> &str

Plus sign

In some locales, this includes left-to-right and other marks. However, the calling code expects it to always only return one glyph.

fn minus_sign(&self) -> &str

Minus sign

In some locales, this includes left-to-right and other marks. However, the calling code expects it to always only return one glyph.

fn engineering_exponent_separator(&self) -> &str

Engineering exponent separator

The symbol or string used to introduce decimal exponent in engineering notation. Defaults to ‘e’.

fn common_exponent_separator(&self) -> &str

Common exponent separator

The symbol sequence used to introduce decimal exponent in “common” (e.g. mathematical or physical) notation. Defaults to ‘×10’. In locales using decimal digits, the exponent after this uses superscript digits.

fn infinity(&self) -> &str

Infinity

The symbol used for infinity.

fn not_a_number(&self) -> &str

Not-a-number

The symbol used for not-a-number.