Trait locale::Localize

pub trait Localize: AsLocalize {
    fn locale_fmt(&self,
                  locale: &Locale,
                  fmt: &str,
                  out: &mut Formatter)
                  -> Result;

    fn localize<'a>(&'a self, fmt: &'a str) -> Localized<'a> { ... }
    fn localize_to<'a>(&'a self,
                       locale: &'a Locale,
                       fmt: &'a str)
                       -> Localized<'a> { ... }
    fn to_local_string(&self, fmt: &str) -> String { ... }
    fn to_locale_string(&self, locale: &Locale, fmt: &str) -> String { ... }
}

Locale-aware formatting for type

Various kins of information like numbers, dates and times, money amounts or physical quantities are written differently depending on language and region. This trait is a main unified interface for formatting of values according to appropriate rules for their type and selected locale.

All the formatting methods take a fmt string argument. This is interpreted in a type-specific way.

For numbers, see numeric

Examples

Simple formatting to string:

Locale::set_current(Locale::new("de-DE").unwrap());
assert_eq!("2,11", 2.11.to_local_string(""));

Required Methods

fn locale_fmt(&self, locale: &Locale, fmt: &str, out: &mut Formatter) -> Result

Core locale-aware formatting function.

This function is for the actual implementation. For regular use, one of the wrappers localize, localize_to, to_local_string or to_locale_string.

Parameters

locale: The locale to format for.

fmt: Format string. Interpretation is format-specific. See trait documentation.

out: Result shall be written to this formatter.

Returns

Simple and totally informationless standard formatting Result.

Panics

Might panic if the underlying type methods return invalid results, for example if a number's Display returns string with characters not acceptable in a number.

In release mode, it should not panic on invalid format, but should just print either nothing or use the default format.

In debug mode it might panic on invalid format to provide better context for debugging (especially since the std::fmt::Result does not provide place for passing any error description).

Provided Methods

fn localize<'a>(&'a self, fmt: &'a str) -> Localized<'a>

Localize for the current locale to standard formatter.

This function is intended for writing to things that accept std::fmt::Display values like standard output or files.

Parameters

fmt: Format string. Interpretation is format-specific. See trait documentation.

Returns

Wrapper implementing std::fmt::Display.

Panics

See locale_fmt.

fn localize_to<'a>(&'a self, locale: &'a Locale, fmt: &'a str) -> Localized<'a>

Localize for specified locale to standard formatter.

This function is intended for writing to things that accept std::fmt::Display values like standard output or files.

Parameters

locale: The locale to format for.

fmt: Format string. Interpretation is format-specific. See trait documentation.

Returns

Wrapper implementing std::fmt::Display.

Panics

See locale_fmt.

fn to_local_string(&self, fmt: &str) -> String

Localize for the current locale to string.

This function is intended for storing the localized form. When writing out, the localize function is probably more efficient as it might avoid some allocations.

Parameters

fmt: Format string. Interpretation is format-specific. See trait documentation.

Returns

Wrapper implementing std::fmt::Display.

Panics

See locale_fmt.

fn to_locale_string(&self, locale: &Locale, fmt: &str) -> String

Localize for specified locale to string.

This function is intended for storing the localized form. When writing out, the localize function is probably more efficient as it might avoid some allocations.

Parameters

locale: The locale to format for.

fmt: Format string. Interpretation is format-specific. See trait documentation.

Returns

Wrapper implementing std::fmt::Display.

Panics

See locale_fmt.